% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_open_library_api.R
\name{call_open_library_api}
\alias{call_open_library_api}
\title{Appel à l'API Open Library}
\usage{
call_open_library_api(root_api = "https://openlibrary.org/api/books", isbn10)
}
\arguments{
\item{root_api}{l'URL de base de l'API}

\item{isbn10}{le numéro ISBN 10}
}
\value{
un tibble avec les informations de l'API
}
\description{
Appel à l'API Open Library
}
\examples{
call_open_library_api(isbn10 = "2365772013")
}
