% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_html_for_modal_modify_book_in_collection.R
\name{create_html_for_modal_modify_book_in_collection}
\alias{create_html_for_modal_modify_book_in_collection}
\title{HTML for modal modify book in collection}
\usage{
create_html_for_modal_modify_book_in_collection(current_book, ns)
}
\arguments{
\item{current_book}{the current book}

\item{ns}{the namespace}
}
\value{
a div with the HTML for the modal to modify a book in the collection
}
\description{
HTML for modal modify book in collection
}
