% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_img_from_dataurl.R
\name{export_img_from_dataurl}
\alias{export_img_from_dataurl}
\title{Export image from data URL}
\usage{
export_img_from_dataurl(dataurl, output_img)
}
\arguments{
\item{dataurl}{A string with the data URL}

\item{output_img}{A string with the path to the output image}
}
\value{
write an img in inst/
}
\description{
Export image from data URL
}
