% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nstars_from_selectinput.R
\name{extract_nstars_from_selectinput}
\alias{extract_nstars_from_selectinput}
\title{Extract n stars from selectinput}
\usage{
extract_nstars_from_selectinput(selectinput_result)
}
\arguments{
\item{selectinput_result}{the result of a selectinput}
}
\value{
the number of stars checked
}
\description{
Extract n stars from selectinput
}
