% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mod_110.R
\name{create_html_for_modal_api_search_result}
\alias{create_html_for_modal_api_search_result}
\title{A function to create the HTML for the modal API search result}
\usage{
create_html_for_modal_api_search_result(book, book_cover)
}
\arguments{
\item{book}{A list with the book information}

\item{book_cover}{A character with the URL of the book cover}
}
\value{
A shiny.tag.list
}
\description{
A function to create the HTML for the modal API search result
}
