% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mod_130.R
\name{prepare_comics_db_to_see_collection}
\alias{prepare_comics_db_to_see_collection}
\title{Prepare comics database to see collection}
\usage{
prepare_comics_db_to_see_collection(comics_db, ns)
}
\arguments{
\item{comics_db}{a comics database}

\item{ns}{the namespace}
}
\value{
a comics database with only the columns needed to see the collection
}
\description{
Prepare comics database to see collection
}
