% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_webcam_utils.R
\name{take_picture_from_webcam_ui}
\alias{take_picture_from_webcam_ui}
\title{Generate a UI for taking a picture from the webcam}
\usage{
take_picture_from_webcam_ui(id)
}
\value{
a shiny tagList
}
\description{
Generate a UI for taking a picture from the webcam
}
