% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cover.R
\name{get_cover}
\alias{get_cover}
\title{Get the cover of a book from Open Library}
\usage{
get_cover(
  root_api = "http://covers.openlibrary.org/b/isbn",
  isbn_number,
  cover_size = "M"
)
}
\arguments{
\item{root_api}{the base URL of the API}

\item{isbn_number}{the ISBN 10 number or the ISBN 13 number of the book}

\item{cover_size}{the size of the cover}
}
\value{
a URL to the cover of the book
}
\description{
Get the cover of a book from Open Library
}
\examples{
get_cover(isbn_number = "9782365772013")
}
