% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_me_one_random_isbn.R
\name{give_me_one_random_isbn}
\alias{give_me_one_random_isbn}
\title{Return a random ISBN from mycomicslibrary::isbn_sample}
\usage{
give_me_one_random_isbn()
}
\value{
a random ISBN
}
\description{
Return a random ISBN from mycomicslibrary::isbn_sample
}
\examples{
give_me_one_random_isbn()
}
