% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stars_rating_div.R
\name{make_stars_rating_div}
\alias{make_stars_rating_div}
\title{Title}
\usage{
make_stars_rating_div(note, max_note = 5)
}
\description{
Description
}
\examples{
make_stars_rating_div(note = 1)
}
