% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comics_db_to_see_wishlist.R
\name{prepare_comics_db_to_see_wishlist}
\alias{prepare_comics_db_to_see_wishlist}
\title{prepare comics database to see wishlist}
\usage{
prepare_comics_db_to_see_wishlist(comics_db, ns)
}
\value{
a comics database with only the columns needed to see the wishlist
}
\description{
prepare comics database to see wishlist
}
