% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_html_for_modal_modify_book_in_collection.R
\name{create_html_for_modal_modify_book_in_collection}
\alias{create_html_for_modal_modify_book_in_collection}
\title{Title}
\usage{
create_html_for_modal_modify_book_in_collection(current_book, ns)
}
\description{
Description
}
\examples{
create_html_for_modal_modify_book_in_collection()
}
