% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mod_130.R
\name{make_stars_rating_div}
\alias{make_stars_rating_div}
\alias{prepare_comics_db_to_see_collection}
\title{Make stars rating div}
\usage{
make_stars_rating_div(note, max_note = 5)

prepare_comics_db_to_see_collection(comics_db, ns)
}
\arguments{
\item{note}{a note between 0 and max_note}

\item{max_note}{the maximum note}

\item{comics_db}{a comics database}

\item{ns}{the namespace}
}
\value{
a div with stars

a comics database with only the columns needed to see the collection
}
\description{
Make stars rating div

Prepare comics database to see collection
}
\examples{
make_stars_rating_div(note = 1)
}
