% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_open_library.R
\name{call_open_library_api}
\alias{call_open_library_api}
\alias{clean_open_library_result}
\title{Appel à l'API Open Library}
\usage{
call_open_library_api(
  root_api = "https://openlibrary.org/api/books",
  isbn_number
)

clean_open_library_result(book_tibble)
}
\arguments{
\item{root_api}{l'URL de base de l'API}

\item{isbn_number}{le numéro ISBN 10}

\item{book_tibble}{a tibble with the result of the API call}
}
\value{
un tibble avec les informations de l'API

a tibble with the cleaned result of the API call
}
\description{
Appel à l'API Open Library

Nettoyage du résultat de l'API Open Library
}
\examples{
call_open_library_api(isbn_number = "2365772013")
isbn_number <- "2365772013"
result <- call_open_library_api(isbn_number = isbn_number)

clean_open_library_result(result)
}
