% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_open_library.R
\name{call_open_library_api}
\alias{call_open_library_api}
\alias{clean_open_library_result}
\alias{get_cover}
\alias{give_me_one_random_isbn}
\title{Appel à l'API Open Library}
\usage{
call_open_library_api(
  root_api = "https://openlibrary.org/api/books",
  isbn_number
)

clean_open_library_result(book_tibble)

get_cover(
  root_api = "http://covers.openlibrary.org/b/isbn",
  isbn_number,
  cover_size = "M",
  resourcepath = resourcePaths()["covers"]
)

give_me_one_random_isbn()
}
\arguments{
\item{root_api}{the base URL of the API}

\item{isbn_number}{the ISBN 10 number or the ISBN 13 number of the book}

\item{book_tibble}{a tibble with the result of the API call}

\item{cover_size}{the size of the cover}

\item{resourcepath}{the path to the resource}
}
\value{
un tibble avec les informations de l'API

a tibble with the cleaned result of the API call

a URL to the cover of the book

a random ISBN
}
\description{
Appel à l'API Open Library

Nettoyage du résultat de l'API Open Library

Get the cover of a book from Open Library

Return a random ISBN from mycomicslibrary::isbn_sample
}
\examples{
call_open_library_api(isbn_number = "2365772013")
isbn_number <- "2365772013"
result <- call_open_library_api(isbn_number = isbn_number)

clean_open_library_result(result)
isbn_number <- "2365772013"
result <- call_open_library_api(isbn_number = isbn_number)

clean_open_library_result(result)
get_cover(isbn_number = "9782365772013")
path <- "home/yohann/Documents/perso/mycomicslibrary/inst/app/www/cover_tmp/9782365772013.jpg"
get_cover(isbn_number = "9782365772013")
path <- "home/yohann/Documents/perso/mycomicslibrary/inst/app/www/cover_tmp/9782365772013.jpg"
give_me_one_random_isbn()
}
