% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_isbn_sample.R
\docType{data}
\name{isbn_sample}
\alias{isbn_sample}
\title{isbn_sample}
\format{
A vector of 50 isbn from my personal collection.
\describe{
}
}
\source{
Source
}
\usage{
isbn_sample
}
\description{
A sample of isbn
}
\keyword{datasets}
