% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_isbn_sample.R
\docType{data}
\name{isbn_sample}
\alias{isbn_sample}
\title{isbn_sample}
\format{
A data frame with  rows and  variables:
\describe{
}
}
\source{
Source
}
\usage{
isbn_sample
}
\description{
Description.
}
\keyword{datasets}
