% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_isbn_sample.R
\docType{data}
\name{isbn_sample}
\alias{isbn_sample}
\title{isbn_sample}
\format{
A data frame with 100000 rows and 2 variables:
\describe{
  \item{ code }{  integer64 }
  \item{ gtinType }{  character }
}
}
\source{
Source
}
\usage{
isbn_sample
}
\description{
Description.
}
\keyword{datasets}
