% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stars_rating_div.R
\name{make_stars_rating_div}
\alias{make_stars_rating_div}
\title{Make stars rating div}
\usage{
make_stars_rating_div(note, max_note = 5)
}
\arguments{
\item{note}{a note between 0 and max_note}

\item{max_note}{the maximum note}
}
\value{
a div with stars
}
\description{
Make stars rating div
}
\examples{
make_stars_rating_div(note = 1)
}
